requirements.txt


python-dotenv

pandas

requests

plotly
